+++ cadc-tap-schema/build.gradle
@@ -15,7 +15,7 @@ sourceCompatibility = 11
 
 group = 'org.opencadc'
 
-version = '1.2.17'
+version = '1.2.18'
 
 description = 'OpenCADC TAP-1.1 tap schema server library'
 def git_url = 'https://github.com/opencadc/tap'
@@ -30,7 +30,7 @@ dependencies {
     implementation 'org.opencadc:cadc-gms:[1.0,2.0)'
     implementation 'org.opencadc:cadc-rest:[1.3.1,2.0)'
     api 'org.opencadc:cadc-dali:[1.2.30,2.0)'
-    api 'org.opencadc:cadc-tap:[1.1.25,2.0)'
+    api 'org.opencadc:cadc-tap:[1.1.26,2.0)'
     
     implementation 'uk.ac.starlink:jcdf:[1.2.3,2.0)'
     implementation 'uk.ac.starlink:stil:[4.0,5.0)'
