+++ cadc-tap/src/main/resources/ucd-list-deprecated.txt
