+++ youcat/build.gradle
@@ -30,8 +30,8 @@ dependencies {
     implementation 'org.opencadc:cadc-adql:[1.1.14,)'
     implementation 'org.opencadc:cadc-uws:[1.1,)'
     implementation 'org.opencadc:cadc-uws-server:[1.3,)'
-    implementation 'org.opencadc:cadc-tap:[1.1.25,)'
-    implementation 'org.opencadc:cadc-tap-schema:[1.2.16,)'
+    implementation 'org.opencadc:cadc-tap:[1.1.26,)'
+    implementation 'org.opencadc:cadc-tap-schema:[1.2.18,)'
     implementation 'org.opencadc:cadc-tap-server:[1.1.38,)'
     implementation 'org.opencadc:cadc-tap-server-pg:[1.1.5,)'
     implementation 'org.opencadc:cadc-adql:[1.1.4,)'
